Split build graph lowering
